+++ src/main/java/com/jf/PetApp/application/investment/dto/InvestmentLotDTO.java
@@ -2,17 +2,18 @@ package com.jf.PetApp.application.investment.dto;
 
 import com.jf.PetApp.core.domain.enums.InvestmentType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record InvestmentLotDTO(
     Integer id,
     String name,
     InvestmentType type,
-    Double quantity,
-    Double purchasePrice,
+    BigDecimal quantity,
+    BigDecimal purchasePrice,
     LocalDate purchaseDate,
-    Double currentPrice,
-    Double investedValue,
-    Double currentValue
+    BigDecimal currentPrice,
+    BigDecimal investedValue,
+    BigDecimal currentValue
 ) {
 }
