+++ src/main/java/com/jf/PetApp/infrastructure/controller/investment/dto/AssetRegistrationDto.java
@@ -6,12 +6,13 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record AssetRegistrationDto(
         @NotBlank(message = "Asset name is required") String name,
-        @NotNull(message = "Quantity is required") @Positive(message = "Quantity must be greater than zero") Double quantity,
-        @NotNull(message = "Purchase price is required") @Positive(message = "Purchase price must be greater than zero") Double purchasePrice,
+        @NotNull(message = "Quantity is required") @Positive(message = "Quantity must be greater than zero") BigDecimal quantity,
+        @NotNull(message = "Purchase price is required") @Positive(message = "Purchase price must be greater than zero") BigDecimal purchasePrice,
         @NotNull(message = "Purchase date is required") LocalDate purchaseDate,
         @NotNull(message = "Asset type is required") InvestmentType type
 ) {
